requirements.txt
scipy
opencv-python
matplotlib
segmentation-models-pytorch
pytesseract
numpy
pywavelets
